server/requirements.txt
# CORE
Django>=5.2.1,<6
djangorestframework>=3.16,<4
psycopg2>=2.9,<3
django-redis>=5.2.1,<6
redis>=5.0.0,<6
djangorestframework-simplejwt>=5.5.0,<6
# UTILS
dotenv>=0.9,<1
django-cors-headers>=4.7.0,<5
pillow>=11.2,<12
django-cleanup>=9.0.0,<10
drf-spectacular>=0.28,<1
django-filter>=25.1,<26
Faker>=18.0.0,<27.0.0 # Added Faker for data seeding
# CELERY
celery>=5.5.2,<5.6
django-celery-beat>=2.8.1,<3
requests>=2.31.0,<3
# WEB3
web3>=7.10.0,<8
